Exception handling

+++ src/main/java/com/DsCommerce/service/ProductService.java
@@ -2,13 +2,22 @@ package com.DsCommerce.service;
 
 import com.DsCommerce.dto.ProductDTO;
 import com.DsCommerce.entities.Product;
+import com.DsCommerce.exceptions.DatabaseException;
+import com.DsCommerce.exceptions.ResourceNotFoundException;
 import com.DsCommerce.repository.ProductRepository;
+import com.fasterxml.jackson.databind.exc.InvalidDefinitionException;
+import jakarta.persistence.EntityNotFoundException;
+import org.apache.coyote.Response;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
+import javax.swing.text.html.parser.Entity;
 import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
@@ -21,7 +30,8 @@ public class ProductService {
 
     @Transactional(readOnly = true)
     public ProductDTO findById(Long id) {
-        Product product = productRepository.findById(id).orElse(null);
+        Product product = productRepository.findById(id).orElseThrow(
+                () -> new ResourceNotFoundException("Recurso não encontrado"));
         return new ProductDTO(product);
     }
 
@@ -44,18 +54,30 @@ public class ProductService {
 
     @Transactional
     public ProductDTO update(Long id, ProductDTO productDTO) {
-        Product entity = productRepository.getReferenceById(id);
+        try {
+            Product entity = productRepository.getReferenceById(id);
 
-        copyDtoToEntity(productDTO, entity);
+            copyDtoToEntity(productDTO, entity);
 
-        entity = productRepository.save(entity);
+            entity = productRepository.save(entity);
 
-        return new ProductDTO(entity);
+            return new ProductDTO(entity);
+        } catch (EntityNotFoundException e) {
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
     }
 
-    @Transactional
+    @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id) {
-        productRepository.deleteById(id);
+        if (!productRepository.existsById(id)) {
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
+        try {
+            productRepository.deleteById(id);
+        }
+        catch (DataIntegrityViolationException e) {
+            throw new DatabaseException("Falha de integridade referencial");
+        }
     }
 
     public void copyDtoToEntity(ProductDTO productDTO, Product entity) {
